+++ src/Fluent.Calculations.DotNetGraph/IDotGraphWriterFile.cs
@@ -1,0 +1,9 @@
+﻿using DotNetGraph.Core;
+
+namespace Fluent.Calculations.DotNetGraph
+{
+    internal interface IDotGraphWriterFile
+    {
+        Task SaveToDot(DotGraph graph, string outputFilePath);
+    }
+}
